+++ app/src/main/java/com/brandonjamesyoung/levelup/constants/Colors.kt
@@ -1,0 +1,13 @@
+package com.brandonjamesyoung.levelup.constants
+
+const val EASY_COLOR: Long = 0xFF09FF00
+
+const val MEDIUM_COLOR: Long = 0xFFFFFF00
+
+const val HARD_COLOR: Long = 0xFFFF0000
+
+const val EXPERT_COLOR: Long = 0xFFDD00FF
+
+const val TEXT_COLOR_PRIMARY: Long = 0xFFFFFFFF
+
+const val TEXT_COLOR_SECONDARY: Long = 0xFF000000