+++ app/src/main/java/com/brandonjamesyoung/levelup/data/QuestHistoryDao.kt
@@ -5,8 +5,9 @@ import kotlinx.coroutines.flow.Flow
 
 @Dao
 interface QuestHistoryDao {
+    @Transaction
     @Query("SELECT * FROM CompletedQuest")
-    fun observeAll(): Flow<List<CompletedQuest>>
+    fun observeAll(): Flow<List<CompletedQuestWithIcon>>
 
     @Insert
     suspend fun insert(completedQuest: CompletedQuest)
@@ -17,7 +18,7 @@ interface QuestHistoryDao {
     @Query("UPDATE CompletedQuest SET iconId = null WHERE iconId IN (:iconIds)")
     suspend fun clearDeletedIcons(iconIds: Set<Int>)
 
-    @Query("UPDATE Quest SET iconId = null")
+    @Query("UPDATE ActiveQuest SET iconId = null")
     suspend fun clearAllIcons()
 
     @Query("SELECT COUNT(*) FROM CompletedQuest")