+++ app/src/main/java/com/brandonjamesyoung/levelup/fragments/QuestHistory.kt
@@ -4,30 +4,27 @@ import android.os.Bundle
 import android.util.Log
 import android.view.View
 import android.widget.Button
-import android.widget.LinearLayout
 import android.widget.TextView
+import androidx.compose.ui.platform.ComposeView
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.activityViewModels
 import androidx.lifecycle.lifecycleScope
 import androidx.navigation.fragment.findNavController
 import com.brandonjamesyoung.levelup.R
-import com.brandonjamesyoung.levelup.data.CompletedQuest
-import com.brandonjamesyoung.levelup.utility.ButtonConverter
-import com.brandonjamesyoung.levelup.utility.CardGenerator
-import com.brandonjamesyoung.levelup.utility.ColorHelper.Companion.darken
+import com.brandonjamesyoung.levelup.data.CompletedQuestWithIcon
+import com.brandonjamesyoung.levelup.data.QuestCard
+import com.brandonjamesyoung.levelup.utility.CardGridCreator
 import com.brandonjamesyoung.levelup.viewmodels.QuestHistoryViewModel
-import com.brandonjamesyoung.levelup.views.QuestCardView
 import dagger.hilt.android.AndroidEntryPoint
 import kotlinx.coroutines.launch
 import javax.inject.Inject
+import kotlin.collections.map
 
 @AndroidEntryPoint
 class QuestHistory: Fragment(R.layout.quest_history) {
     private val viewModel: QuestHistoryViewModel by activityViewModels()
 
-    @Inject lateinit var cardGenerator: CardGenerator
-
-    @Inject lateinit var buttonConverter: ButtonConverter
+    @Inject lateinit var cardCreator: CardGridCreator
 
     private fun navigateToQuestList() {
         findNavController().navigate(R.id.action_questHistory_to_questList)
@@ -43,31 +40,6 @@ class QuestHistory: Fragment(R.layout.quest_history) {
         }
     }
 
-    // TODO similar to createQuestCard() in QuestList.kt, possible simplification opportunity
-    private fun createQuestHistoryCard(completedQuest: CompletedQuest) : QuestCardView {
-        val view = requireView()
-        val questName = completedQuest.name ?: getString(R.string.placeholder_text)
-
-        val difficultyColorId = cardGenerator.difficultyColorMap[completedQuest.difficulty]
-            ?: throw IllegalArgumentException("Given card difficulty is not a valid value.")
-
-        val theme = view.context.theme
-        val difficultyColor: Int = resources.getColor(difficultyColorId, theme)
-        val backgroundColor: Int = darken(difficultyColor, DARKNESS_FACTOR)
-
-        val newCard = cardGenerator.createSimpleCard(
-            name = questName,
-            backgroundColorInt = backgroundColor,
-            iconId = completedQuest.iconId,
-            view = view,
-            iconReader = viewModel,
-            lifecycleScope = lifecycleScope
-        )
-
-        newCard.disableClickableIcon()
-        return newCard
-    }
-
     private fun showNoQuestsMessage() {
         val view = requireView()
         val noIconsMessage = view.findViewById<TextView>(R.id.NoQuestsMessage)
@@ -80,19 +52,21 @@ class QuestHistory: Fragment(R.layout.quest_history) {
         noIconsMessage.visibility = View.GONE
     }
 
+    // Add quest cards to lazy vertical grid
+    private fun reloadLazyQuestGrid(quests: List<CompletedQuestWithIcon>) {
+        if (quests.isEmpty()) showNoQuestsMessage() else hideNoQuestsMessage()
+        val sortedQuests = quests.sortedByDescending { it.completedQuest.dateCompleted }
+        val cards: List<QuestCard> = sortedQuests.map { QuestCard(it.completedQuest, it.icon) }
+        val composeView = requireView().findViewById<ComposeView>(R.id.QuestHistoryComposeView)
+
+        composeView.setContent {
+            cardCreator.QuestGridView(cards)
+        }
+    }
+
     private fun setupObservables() {
-        viewModel.questHistoryList.observe(viewLifecycleOwner) { questList ->
-            val view = requireView()
-            val questListLayout = view.findViewById<LinearLayout>(R.id.QuestHistoryLinearLayout)
-            questListLayout.removeAllViews()
-            val sortedQuestHistory = questList.sortedByDescending { it.dateCompleted }
-
-            for (completedQuest in sortedQuestHistory) {
-                val newCard = createQuestHistoryCard(completedQuest)
-                questListLayout.addView(newCard)
-            }
-
-            if (questList.isEmpty()) showNoQuestsMessage() else hideNoQuestsMessage()
+        viewModel.questHistoryList.observe(viewLifecycleOwner) {
+            reloadLazyQuestGrid(it)
         }
     }